abm users

+++ src/pages/sysAdmin/Users.jsx
@@ -1,19 +1,94 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
+import { AuthenticationContext } from '../../components/services/authentication/UserAuthenticationContext';
 
-function Users({ getAllUsers, users, setUsers, currentUser }) {
+const styles = {
+  container: {
+    marginRight: '400px',
+    marginLeft:'60px',
+    marginTop:'60px',
+    marginBottom:'200px',
+    fontFamily: 'Arial, sans-serif',
+  },
+  header: {
+    textAlign: 'left',
+    color: '#333',
+    marginLeft: '10px',
+  },
+  form: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'flex-start',
+    marginBottom: '20px',
+    marginLeft: '10px',
+  },
+  input: {
+    margin: '5px',
+    padding: '10px',
+    borderRadius: '5px',
+    border: '1px solid #ccc',
+  },
+  label: {
+    margin: '5px',
+    fontWeight: 'bold',
+  },
+  checkbox: {
+    marginLeft: '10px',
+  },
+  button: {
+    margin: '5px',
+    padding: '10px 20px',
+    borderRadius: '5px',
+    border: 'none',
+    backgroundColor: '#28a745',
+    color: 'white',
+    cursor: 'pointer',
+  },
+  redButton: {
+    margin: '5px',
+    padding: '10px 20px',
+    borderRadius: '5px',
+    border: 'none',
+    backgroundColor: '#dc3545',
+    color: 'white',
+    cursor: 'pointer',
+  },
+  list: {
+    listStyleType: 'none',
+    padding: '0',
+  },
+  listItem: {
+    marginBottom: '10px',
+  },
+  userInfo: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    padding: '10px',
+    border: '1px solid #ccc',
+    borderRadius: '5px',
+  },
+  userActions: {
+    display: 'flex',
+    gap: '10px',
+  }
+};
+
+function Users({ getAllUsers, users, setUsers}) {
+  const { currentUser } = useContext(AuthenticationContext);
   const [newUser, setNewUser] = useState({
     firstName: '',
     lastName: '',
     email: '',
     password: '',
-    status: false, 
-    isSysAdmin: false, // se pone asi si no seleccionas
+    status: false,
+    isSysAdmin: false,
   });
 
   const [editUser, setEditUser] = useState(null);
 
   // agregar un nuevo usuario
-  const handleAddUser = async () => {
+  const handleAddUser = async (e) => {
+    e.preventDefault();
     try {
       const response = await fetch("http://localhost:8000/users", {
         method: 'POST',
@@ -33,7 +108,7 @@ function Users({ getAllUsers, users, setUsers, currentUser }) {
           email: '',
           password: '',
           status: false,
-          isSysAdmin: false
+          isSysAdmin: false,
         });
       } else {
         console.error('Error al crear el usuario:', response.statusText);
@@ -45,21 +120,24 @@ function Users({ getAllUsers, users, setUsers, currentUser }) {
 
   // eliminar un usuario
   const handleDeleteUser = async (userId) => {
-    try {
-      const response = await fetch(`http://localhost:8000/users/${userId}`, {
-        method: 'DELETE',
-        headers: {
-          "authorization": `Bearer ${currentUser.accessToken}`
-        },
-      });
+    // Confirmación antes de eliminar
+    if (window.confirm("¿Estás seguro de que deseas eliminar este usuario?")) {
+      try {
+        const response = await fetch(`http://localhost:8000/users/${userId}`, {
+          method: 'DELETE',
+          headers: {
+            "authorization": `Bearer ${currentUser.accessToken}`
+          },
+        });
 
-      if (response.ok) {
-        setUsers(users.filter(user => user.id !== userId));
-      } else {
-        console.error('Error al eliminar el usuario:', response.statusText);
+        if (response.ok) {
+          setUsers(users.filter(user => user.id !== userId));
+        } else {
+          console.error('Error al eliminar el usuario:', response.statusText);
+        }
+      } catch (error) {
+        console.error('Error al eliminar el usuario:', error);
       }
-    } catch (error) {
-      console.error('Error al eliminar el usuario:', error);
     }
   };
 
@@ -87,7 +165,7 @@ function Users({ getAllUsers, users, setUsers, currentUser }) {
     }
   };
 
-  //maneja cambios inputs del formulario de creacion de usuario
+  // maneja cambios inputs del formulario de creacion de usuario
   const handleInputChange = (e) => {
     const { name, value } = e.target;
     if (name === 'status' || name === 'isSysAdmin') {
@@ -97,7 +175,7 @@ function Users({ getAllUsers, users, setUsers, currentUser }) {
     }
   };
 
-  //  maneja cambios en los inputs del formulario de edicion de usuario
+  // maneja cambios en los inputs del formulario de edicion de usuario
   const handleEditInputChange = (e, user) => {
     const { name, value } = e.target;
     if (name === 'status' || name === 'isSysAdmin') {
@@ -112,57 +190,55 @@ function Users({ getAllUsers, users, setUsers, currentUser }) {
     setEditUser(user);
   };
 
-  // para obtener usuarios al cargar el componente
-  useEffect(() => {
-    getAllUsers(); 
-  }, [getAllUsers]);
+ 
 
   return (
-    <div>
-      <h2>Usuarios</h2>
+    <div style={styles.container}>
+      <h2 style={styles.header}>Usuarios</h2>
 
-      <form onSubmit={handleAddUser}>
-        <input type="text" name="firstName" value={newUser.firstName} onChange={handleInputChange} placeholder="Nombre" />
-        <input type="text" name="lastName" value={newUser.lastName} onChange={handleInputChange} placeholder="Apellido" />
-        <input type="email" name="email" value={newUser.email} onChange={handleInputChange} placeholder="Email" />
-        <input type="password" name="password" value={newUser.password} onChange={handleInputChange} placeholder="Contraseña" />
-        <label>
+      <form onSubmit={handleAddUser} style={styles.form}>
+        <input type="text" name="firstName" value={newUser.firstName} onChange={handleInputChange} placeholder="Nombre" style={styles.input} />
+        <input type="text" name="lastName" value={newUser.lastName} onChange={handleInputChange} placeholder="Apellido" style={styles.input} />
+        <input type="email" name="email" value={newUser.email} onChange={handleInputChange} placeholder="Email" style={styles.input} />
+        <input type="password" name="password" value={newUser.password} onChange={handleInputChange} placeholder="Contraseña" style={styles.input} />
+        <label style={styles.label}>
           Administrador:
-          <input type="checkbox" name="status" checked={newUser.status} onChange={handleInputChange} />
+          <input type="checkbox" name="status" checked={newUser.status} onChange={handleInputChange} style={styles.checkbox} />
         </label>
-        <label>
-          Administrador del sistemas:
-          <input type="checkbox" name="isSysAdmin" checked={newUser.isSysAdmin} onChange={handleInputChange} />
+        <label style={styles.label}>
+          Administrador del sistema:
+          <input type="checkbox" name="isSysAdmin" checked={newUser.isSysAdmin} onChange={handleInputChange} style={styles.checkbox} />
         </label>
-        <button type="submit">Agregar Usuario</button>
+        <button type="submit" style={styles.button}>Agregar Usuario</button>
       </form>
 
-    
-      <ul>
+      <ul style={styles.list}>
         {users.map(user => (
-          <li key={user.id}>
+          <li key={user.id} style={styles.listItem}>
             {editUser?.id === user.id ? (
               <div>
-                <input type="text" name="firstName" value={editUser.firstName} onChange={(e) => handleEditInputChange(e, editUser)} />
-                <input type="text" name="lastName" value={editUser.lastName} onChange={(e) => handleEditInputChange(e, editUser)} />
-                <input type="email" name="email" value={editUser.email} onChange={(e) => handleEditInputChange(e, editUser)} />
-                <input type="password" name="password" value={editUser.password} onChange={(e) => handleEditInputChange(e, editUser)} />
-                <label>
+                <input type="text" name="firstName" value={editUser.firstName} onChange={(e) => handleEditInputChange(e, editUser)} style={styles.input} />
+                <input type="text" name="lastName" value={editUser.lastName} onChange={(e) => handleEditInputChange(e, editUser)} style={styles.input} />
+                <input type="email" name="email" value={editUser.email} onChange={(e) => handleEditInputChange(e, editUser)} style={styles.input} />
+                <input type="password" name="password" value={editUser.password} onChange={(e) => handleEditInputChange(e, editUser)} style={styles.input} />
+                <label style={styles.label}>
                   Administrador:
-                  <input type="checkbox" name="status" checked={editUser.status} onChange={(e) => handleEditInputChange(e, editUser)} />
+                  <input type="checkbox" name="status" checked={editUser.status} onChange={(e) => handleEditInputChange(e, editUser)} style={styles.checkbox} />
                 </label>
-                <label>
+                <label style={styles.label}>
                   Administrador del sistema:
-                  <input type="checkbox" name="isSysAdmin" checked={editUser.isSysAdmin} onChange={(e) => handleEditInputChange(e, editUser)} />
+                  <input type="checkbox" name="isSysAdmin" checked={editUser.isSysAdmin} onChange={(e) => handleEditInputChange(e, editUser)} style={styles.checkbox} />
                 </label>
-                <button onClick={handleEditUser}>Guardar</button>
-                <button onClick={() => setEditUser(null)}>Cancelar</button>
+                <button onClick={handleEditUser} style={styles.button}>Guardar</button>
+                <button onClick={() => setEditUser(null)} style={styles.redButton}>Cancelar</button>
               </div>
             ) : (
-              <div>
+              <div style={styles.userInfo}>
                 <span>{user.firstName} {user.lastName}</span>
-                <button onClick={() => toggleEditUser(user)}>Editar</button>
-                <button onClick={() => handleDeleteUser(user.id)}>Eliminar</button>
+                <div style={styles.userActions}>
+                  <button onClick={() => toggleEditUser(user)} style={styles.button}>Editar</button>
+                  <button onClick={() => handleDeleteUser(user.id)} style={styles.redButton}>Eliminar</button>
+                </div>
               </div>
             )}
           </li>
